+++ internal/bufpushaction/testdata/empty_module/buf.yaml
@@ -1,0 +1,2 @@
+version: v1
+name: buf.build/foo/bar
