+++ internal/bufpushaction/testdata/invalid_config/buf.yaml
@@ -1,0 +1,1 @@
+this is not a valid config file
